+++ crates/compiler/tests/fixtures/bundle/export/doubleReexportAll/foo.ts
@@ -1,0 +1,1 @@
+export { Worker } from './reexport';