+++ crates/plugin_runtime/src/lib.rs
@@ -235,8 +235,11 @@ impl Plugin for FarmPluginRuntime {
 
       if module.module_type.is_script() {
         let ast = &module.meta.as_script().ast;
-
-        if farmfe_toolkit::swc_ecma_utils::contains_top_level_await(ast) {
+        let dependencies = module_graph.dependencies(module_id);
+        let is_deps_async = dependencies
+          .iter()
+          .any(|(dep, edge)| async_modules.contains(dep) && !edge.is_dynamic());
+        if is_deps_async || farmfe_toolkit::swc_ecma_utils::contains_top_level_await(ast) {
           added_async_modules.push(module_id.clone());
         }
       }