+++ packages/core/src/config/index.ts
@@ -22,6 +22,7 @@ import {
   colors,
   isArray,
   isEmptyObject,
+  isNodeEnv,
   isObject,
   isWindows,
   normalizePath
@@ -638,7 +639,8 @@ const formatFromExt: Record<string, Format> = {
   cjs: 'cjs',
   mjs: 'esm',
   cts: 'cjs',
-  mts: 'esm'
+  mts: 'esm',
+  js: 'esm'
 };
 
 const formatToExt: Record<Format, string> = {